rename package tool to tools
