+++ lisa/tools/uname.py
